Bug fix andrew 1 
Fixed the merge issues!

+++ src/model/Ship.cs
@@ -69,7 +69,7 @@ namespace Battleship
         {
             _shipName = ship;
             _tiles = new List<Tile>();
-            _sizeOfShip = _shipName;
+            _sizeOfShip = (int)_shipName;
         }
 
         public void AddTile(Tile tile)
